requirements.txt
absl-py==1.3.0
adal==1.2.7
asttokens==2.1.0
astunparse==1.6.3
awscli==1.27.16
azure-common==1.1.28
azure-core==1.26.1
azure-identity==1.12.0
azure-mgmt-core==1.3.2
azure-mgmt-media==10.1.0
azure-storage-blob==12.14.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
boto3==1.26.16
botocore==1.29.16
cachetools==5.2.0
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.4
comm==0.1.2
cryptography==38.0.3
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.2
dnspython==2.1.0
docker==6.0.1
docutils==0.16
email-validator==1.1.2
executing==1.2.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-WTF==0.15.1
flatbuffers==22.11.23
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
imageai==2.1.6
importlib-metadata==5.0.0
install==1.3.5
ipykernel==6.21.2
ipython==8.10.0
isodate==0.6.1
itsdangerous==1.1.0
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==2.11.2
jmespath==1.0.1
jupyter_client==8.0.2
jupyter_core==5.2.0
keras==2.11.0
keras-resnet==0.2.0
keyring==8.7
keyrings.alt==4.2.0
kiwisolver==1.4.4
libclang==14.0.6
mal-api==0.4.0
Markdown==3.4.1
MarkupSafe==1.1.1
matplotlib==3.3.2
matplotlib-inline==0.1.6
mongoengine==0.23.1
more-itertools==9.0.0
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
mysql-connector-python==8.0.32
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pafy==0.5.5
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.0.0
portalocker==2.6.0
prompt-toolkit==3.0.33
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pymongo==4.3.3
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2022.1
pywin32==305
pywin32-ctypes==0.2.0
PyYAML==5.4.1
pyzmq==25.0.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
scipy==1.4.1
six==1.16.0
smmap==5.0.0
soupsieve==2.2.1
stack-data==0.6.1
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.7.2
voluptuous==0.13.1
wcwidth==0.2.5
websocket-client==1.4.2
Werkzeug==1.0.1
wrapt==1.14.1
WTForms==2.3.3
zipp==3.10.0
